backend/requirements.txt
# requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.7
alembic>=1.12.0

# Document Processing
PyMuPDF>=1.23.0
langchain>=0.0.330
langchain-community>=0.0.10

# Vector Search & Embeddings  
weaviate-client>=3.25.0
sentence-transformers>=2.2.2
scikit-learn>=1.3.0
numpy>=1.24.0

# LLM Providers
google-generativeai>=0.3.0
openai>=1.3.0
anthropic>=0.7.0
ollama>=0.1.7

# Utilities
python-dotenv>=1.0.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
httpx>=0.25.0
aiofiles>=23.2.0
redis>=5.0.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.6.0
